Use single files for streams
