+++ tap_github/streams/__init__.py
@@ -1,0 +1,10 @@
+from tap_github.streams.commits import CommitsStream
+from tap_github.streams.community_profile import CommunityProfileStream
+from tap_github.streams.issue_comments import IssueCommentsStream
+from tap_github.streams.issue_events import IssueEventsStream
+from tap_github.streams.issues import IssuesStream
+from tap_github.streams.pull_request_files import PullRequestFilesStream
+from tap_github.streams.pull_requests import PullRequestsStream
+from tap_github.streams.readme import ReadmeStream
+from tap_github.streams.repository import RepositoryStream
+from tap_github.streams.star_gazers import StargazersStream
