+++ tap_github/streams/issue_events.py
@@ -1,0 +1,64 @@
+from typing import Optional, Iterable, Dict, Any
+
+from singer_sdk import typing as th
+
+from tap_github.client import GitHubStream
+from tap_github.streams.repository import RepositoryStream
+
+
+class IssueEventsStream(GitHubStream):
+    """
+    Defines 'IssueEvents' stream.
+    Issue events are fetched from the repository level (as opposed to per issue)
+    to optimize for API quota usage.
+    """
+
+    name = "issue_events"
+    path = "/repos/{org}/{repo}/issues/events"
+    primary_keys = ["id"]
+    replication_key = "created_at"
+    parent_stream_type = RepositoryStream
+    state_partitioning_keys = ["repo", "org"]
+    ignore_parent_replication_key = False
+
+    def get_records(self, context: Optional[dict] = None) -> Iterable[Dict[str, Any]]:
+        """Return a generator of row-type dictionary objects.
+
+        Each row emitted should be a dictionary of property names to their values.
+        """
+        if context and context.get("events", None) == 0:
+            self.logger.debug(f"No events detected. Skipping '{self.name}' sync.")
+            return []
+
+        return super().get_records(context)
+
+    def post_process(self, row: dict, context: Optional[dict] = None) -> dict:
+        row["issue_number"] = int(row["issue"].pop("number"))
+        row["issue_url"] = row["issue"].pop("url")
+        return row
+
+    schema = th.PropertiesList(
+        th.Property("id", th.IntegerType),
+        th.Property("node_id", th.StringType),
+        th.Property("repo", th.StringType),
+        th.Property("org", th.StringType),
+        th.Property("issue_number", th.IntegerType),
+        th.Property("issue_url", th.StringType),
+        th.Property("event", th.StringType),
+        th.Property("commit_id", th.StringType),
+        th.Property("commit_url", th.StringType),
+        th.Property("created_at", th.DateTimeType),
+        th.Property(
+            "actor",
+            th.ObjectType(
+                th.Property("login", th.StringType),
+                th.Property("id", th.IntegerType),
+                th.Property("node_id", th.StringType),
+                th.Property("avatar_url", th.StringType),
+                th.Property("gravatar_id", th.StringType),
+                th.Property("html_url", th.StringType),
+                th.Property("type", th.StringType),
+                th.Property("site_admin", th.BooleanType),
+            ),
+        ),
+    ).to_dict()
