+++ tap_github/streams/pull_requests.py
@@ -1,0 +1,270 @@
+from typing import Optional, Any, Dict
+
+from singer_sdk import typing as th
+
+from tap_github.client import GitHubStream
+from tap_github.streams.repository import RepositoryStream
+
+
+class PullRequestsStream(GitHubStream):
+    """Defines 'PullRequests' stream."""
+
+    name = "pull_requests"
+    path = "/repos/{org}/{repo}/pulls"
+    primary_keys = ["id"]
+    replication_key = "updated_at"
+    parent_stream_type = RepositoryStream
+    ignore_parent_replication_key = False
+    state_partitioning_keys = ["repo", "org"]
+
+    def get_url_params(
+        self, context: Optional[dict], next_page_token: Optional[Any]
+    ) -> Dict[str, Any]:
+        """Return a dictionary of values to be used in URL parameterization."""
+        assert context is not None, f"Context cannot be empty for '{self.name}' stream."
+        params = super().get_url_params(context, next_page_token)
+        # Fetch all pull requests regardless of state (OPEN, CLOSED, MERGED).
+        params["state"] = "all"
+        return params
+
+    def get_child_context(self, record: dict, context: Optional[dict]) -> dict:
+        """
+        Give the PR number to children streams
+        """
+
+        return {
+            **super().get_child_context(record, context),
+            **{
+                "pull_number": record["number"],
+            },
+        }
+
+    @property
+    def http_headers(self) -> dict:
+        """Return the http headers needed.
+
+        Overridden to use beta endpoint which includes reactions as described here:
+        https://developer.github.com/changes/2016-05-12-reactions-api-preview/
+        """
+        headers = super().http_headers
+        headers["Accept"] = "application/vnd.github.squirrel-girl-preview"
+        return headers
+
+    schema = th.PropertiesList(
+        # Parent keys
+        th.Property("repo", th.StringType),
+        th.Property("org", th.StringType),
+        # PR keys
+        th.Property("id", th.IntegerType),
+        th.Property("node_id", th.StringType),
+        th.Property("url", th.StringType),
+        th.Property("html_url", th.StringType),
+        th.Property("diff_url", th.StringType),
+        th.Property("patch_url", th.StringType),
+        th.Property("number", th.IntegerType),
+        th.Property("updated_at", th.DateTimeType),
+        th.Property("created_at", th.DateTimeType),
+        th.Property("closed_at", th.DateTimeType),
+        th.Property("merged_at", th.DateTimeType),
+        th.Property("state", th.StringType),
+        th.Property("title", th.StringType),
+        th.Property("locked", th.BooleanType),
+        th.Property("comments", th.IntegerType),
+        th.Property("author_association", th.StringType),
+        th.Property("body", th.StringType),
+        th.Property("merge_commit_sha", th.StringType),
+        th.Property("draft", th.BooleanType),
+        th.Property("commits_url", th.StringType),
+        th.Property("review_comments_url", th.StringType),
+        th.Property("review_comment_url", th.StringType),
+        th.Property("comments_url", th.StringType),
+        th.Property("statuses_url", th.StringType),
+        th.Property(
+            "user",
+            th.ObjectType(
+                th.Property("login", th.StringType),
+                th.Property("id", th.IntegerType),
+                th.Property("node_id", th.StringType),
+                th.Property("avatar_url", th.StringType),
+                th.Property("gravatar_id", th.StringType),
+                th.Property("html_url", th.StringType),
+                th.Property("type", th.StringType),
+                th.Property("site_admin", th.BooleanType),
+            ),
+        ),
+        th.Property(
+            "labels",
+            th.ArrayType(
+                th.ObjectType(
+                    th.Property("id", th.IntegerType),
+                    th.Property("node_id", th.StringType),
+                    th.Property("url", th.StringType),
+                    th.Property("name", th.StringType),
+                    th.Property("description", th.StringType),
+                    th.Property("color", th.StringType),
+                    th.Property("default", th.BooleanType),
+                ),
+            ),
+        ),
+        th.Property(
+            "reactions",
+            th.ObjectType(
+                th.Property("url", th.StringType),
+                th.Property("total_count", th.IntegerType),
+                th.Property("+1", th.IntegerType),
+                th.Property("-1", th.IntegerType),
+                th.Property("laugh", th.IntegerType),
+                th.Property("hooray", th.IntegerType),
+                th.Property("confused", th.IntegerType),
+                th.Property("heart", th.IntegerType),
+                th.Property("rocket", th.IntegerType),
+                th.Property("eyes", th.IntegerType),
+            ),
+        ),
+        th.Property(
+            "assignee",
+            th.ObjectType(
+                th.Property("login", th.StringType),
+                th.Property("id", th.IntegerType),
+                th.Property("node_id", th.StringType),
+                th.Property("avatar_url", th.StringType),
+                th.Property("gravatar_id", th.StringType),
+                th.Property("html_url", th.StringType),
+                th.Property("type", th.StringType),
+                th.Property("site_admin", th.BooleanType),
+            ),
+        ),
+        th.Property(
+            "assignees",
+            th.ArrayType(
+                th.ObjectType(
+                    th.Property("login", th.StringType),
+                    th.Property("id", th.IntegerType),
+                    th.Property("node_id", th.StringType),
+                    th.Property("avatar_url", th.StringType),
+                    th.Property("gravatar_id", th.StringType),
+                    th.Property("html_url", th.StringType),
+                    th.Property("type", th.StringType),
+                    th.Property("site_admin", th.BooleanType),
+                ),
+            ),
+        ),
+        th.Property(
+            "requested_reviewers",
+            th.ArrayType(
+                th.ObjectType(
+                    th.Property("login", th.StringType),
+                    th.Property("id", th.IntegerType),
+                    th.Property("node_id", th.StringType),
+                    th.Property("avatar_url", th.StringType),
+                    th.Property("gravatar_id", th.StringType),
+                    th.Property("html_url", th.StringType),
+                    th.Property("type", th.StringType),
+                    th.Property("site_admin", th.BooleanType),
+                ),
+            ),
+        ),
+        th.Property(
+            "milestone",
+            th.ObjectType(
+                th.Property("html_url", th.StringType),
+                th.Property("node_id", th.StringType),
+                th.Property("id", th.IntegerType),
+                th.Property("number", th.IntegerType),
+                th.Property("state", th.StringType),
+                th.Property("title", th.StringType),
+                th.Property("description", th.StringType),
+                th.Property(
+                    "creator",
+                    th.ObjectType(
+                        th.Property("login", th.StringType),
+                        th.Property("id", th.IntegerType),
+                        th.Property("node_id", th.StringType),
+                        th.Property("avatar_url", th.StringType),
+                        th.Property("gravatar_id", th.StringType),
+                        th.Property("html_url", th.StringType),
+                        th.Property("type", th.StringType),
+                        th.Property("site_admin", th.BooleanType),
+                    ),
+                ),
+                th.Property("open_issues", th.IntegerType),
+                th.Property("closed_issues", th.IntegerType),
+                th.Property("created_at", th.DateTimeType),
+                th.Property("updated_at", th.DateTimeType),
+                th.Property("closed_at", th.DateTimeType),
+                th.Property("due_on", th.DateTimeType),
+            ),
+        ),
+        th.Property("locked", th.BooleanType),
+        th.Property(
+            "pull_request",
+            th.ObjectType(
+                th.Property("html_url", th.StringType),
+                th.Property("url", th.StringType),
+                th.Property("diff_url", th.StringType),
+                th.Property("patch_url", th.StringType),
+            ),
+        ),
+        th.Property(
+            "head",
+            th.ObjectType(
+                th.Property("label", th.StringType),
+                th.Property("ref", th.StringType),
+                th.Property("sha", th.StringType),
+                th.Property(
+                    "user",
+                    th.ObjectType(
+                        th.Property("login", th.StringType),
+                        th.Property("id", th.IntegerType),
+                        th.Property("node_id", th.StringType),
+                        th.Property("avatar_url", th.StringType),
+                        th.Property("gravatar_id", th.StringType),
+                        th.Property("html_url", th.StringType),
+                        th.Property("type", th.StringType),
+                        th.Property("site_admin", th.BooleanType),
+                    ),
+                ),
+                th.Property(
+                    "repo",
+                    th.ObjectType(
+                        th.Property("id", th.IntegerType),
+                        th.Property("node_id", th.StringType),
+                        th.Property("name", th.StringType),
+                        th.Property("full_name", th.StringType),
+                        th.Property("html_url", th.StringType),
+                    ),
+                ),
+            ),
+        ),
+        th.Property(
+            "base",
+            th.ObjectType(
+                th.Property("label", th.StringType),
+                th.Property("ref", th.StringType),
+                th.Property("sha", th.StringType),
+                th.Property(
+                    "user",
+                    th.ObjectType(
+                        th.Property("login", th.StringType),
+                        th.Property("id", th.IntegerType),
+                        th.Property("node_id", th.StringType),
+                        th.Property("avatar_url", th.StringType),
+                        th.Property("gravatar_id", th.StringType),
+                        th.Property("html_url", th.StringType),
+                        th.Property("type", th.StringType),
+                        th.Property("site_admin", th.BooleanType),
+                    ),
+                ),
+                th.Property(
+                    "repo",
+                    th.ObjectType(
+                        th.Property("id", th.IntegerType),
+                        th.Property("node_id", th.StringType),
+                        th.Property("name", th.StringType),
+                        th.Property("full_name", th.StringType),
+                        th.Property("html_url", th.StringType),
+                    ),
+                ),
+            ),
+        ),
+    ).to_dict()
