+++ tap_github/streams/readme.py
@@ -1,0 +1,39 @@
+from singer_sdk import typing as th
+
+from tap_github.client import GitHubStream
+from tap_github.streams.repository import RepositoryStream
+
+
+class ReadmeStream(GitHubStream):
+    name = "readme"
+    path = "/repos/{org}/{repo}/readme"
+    primary_keys = ["repo", "org"]
+    parent_stream_type = RepositoryStream
+    ignore_parent_replication_key = False
+    state_partitioning_keys = ["repo", "org"]
+
+    schema = th.PropertiesList(
+        # Parent Keys
+        th.Property("repo", th.StringType),
+        th.Property("org", th.StringType),
+        # README Keys
+        th.Property("type", th.StringType),
+        th.Property("encoding", th.StringType),
+        th.Property("size", th.IntegerType),
+        th.Property("name", th.StringType),
+        th.Property("path", th.StringType),
+        th.Property("content", th.StringType),
+        th.Property("sha", th.StringType),
+        th.Property("url", th.StringType),
+        th.Property("git_url", th.StringType),
+        th.Property("html_url", th.StringType),
+        th.Property("download_url", th.StringType),
+        th.Property(
+            "_links",
+            th.ObjectType(
+                th.Property("git", th.StringType),
+                th.Property("self", th.StringType),
+                th.Property("html", th.StringType),
+            ),
+        ),
+    ).to_dict()
